#279 Merged the test case files into a single file
